+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Short alias `/m` for `/moo` command
   - Copy inventories to clipboard as JSON with <kbd>CTRL</kbd> + <kbd>C</kbs>
 - Added sound when a best friend comes online (deactivated by default)
+- Check how long current world has been loaded
+  - ≈ when server was last restarted
+  - via command `/moo worldage`
+  - notification when joining a recently loaded or a very old server (toggleable via config)
 
 ### Changed
 - Item age: show timestamp in the local timezone instead of "SkyBlock"-timezone (Eastern Time; also fixed the incorrect 12h ↔ 24h clock conversion)